Altered the base way Stacks are structured.
Now there is a special "head" of the stack using a Stack struct. This
contains the first Node of the stack as well as the size of the key used
for this stack. This means that stacks now accept any kind of key. All
the functions are being rewritten to accomodate for this. Currently, all
stack functions except stack_remove are working.