Tried to get basic movement to work.
Finished moving hashmap to a submodule.
Added new function move_entity in explore_entity to handle moving
entities from one cell in the grid to the other. Added new parameters to
the game_state and renamed columns to cols to save space.
Still trying to figure out why the ships are artifacting. Also added
rudimentary collision detection for the borders, but it should handle
new Entities in play too.